Nhse o34 d34.upstream (#8)
* Use erl format (#129)

* Use erl format

* Add format check to GHA

* Fix eqwalizer errors (#130)

* Use erl format

* Fix eqwalizer errors

Specific ignores used in some cases

* Fix eqwalizer errors

Specific ignores used in some cases

* Allow for lack of dynamic() support pre OTP 26

* Mas d34 orkv.i42 cache filter (#132)

* Add basic support for key_filter

Allow for key_filter to exclude keys from exchanges and tree caches.

the filter is applied in three places in the controller:
- when calling aae_put;
- when calling rebuild trees;
- when claling fetch_clocks (Specifically for the tree cache updates)

For th e exchange the filter is applied before compare_clocks.  It is assumed that results from folds include filtered keys, and they are then filtered before comparison.  This allows for exchanges without a filter (e.g. when doing an exchange for the specific bucket)

* Initial test using mock vnode

* Provide a pretty-printed count of buckets - before comparison

Helpful when troubleshooting

* Extend the testing of Bucket filters

* Add new option to type

* Change name of key_filter_fun() type

Should be key_include_fun() to make it explicit that the key is included if response is true.

* Update README for OpenRiak

* Refer to 3.4 not 3.2

* Update meck reference - so can eqc test in OTP 24